Removed controlled component and simply render info from neighborhood GeoJSON below map in plain JSX.
