+++ src/LeafletPlusData/CustomControl.js
